# swagger.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: 360 Review Client
paths:
...
  '/api/v1/auth/email/verification':
    post:
      tags:
        - Auth
      summary: Подтверждение email'а
      operationId: verifyEmail
      requestBody:
        $ref: '#/components/requestBodies/VerifyEmailRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseToken'
        '404':
          description: User or code not found
        '422':
          description: Validation Error
  '/api/v1/auth/password/reset':
    post:
      tags:
        - Auth
      summary: Сброс пароля
      operationId: resetPassword
      requestBody:
        $ref: '#/components/requestBodies/ResetPasswordRequest'
      responses:
        '200':
          description: Password successfully changed
        '404':
          description: User or code not found
        '422':
          description: Validation Error
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    ResponseToken:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
  requestBodies:
...
    VerifyEmailRequest:
      required: true
      content:
        application/x-www-form-urlencoded:
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
              code:
                type: string
            required:
              - email
              - password
              - code
    ResetPasswordRequest:
      required: true
      content:
        application/x-www-form-urlencoded:
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
              code:
                type: string
            required:
              - email
              - password
              - code
...
security:
  - ApiKeyAuth: []